Move Register to containers
